--- src/lib/components/Sidebar/SideBarContainer.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-interface IProps {
-    labelFor: string;
-}
-
-const SideBarContainer = ({ labelFor }: IProps) => {
-    return <></>;
-};
-
-export default SideBarContainer;
