+++ src/lib/components/shared/loaders/index.ts
@@ -1,0 +1,1 @@
+export { Spinner } from './Spinner';
